glibc: Remove the issetugid patch.
It's not safer than get*id.

The Glibc issetugid() patch is no longer used. issetugid() could be preloaded
from a user-defined library, just like getuid() or getgid(), so issetugid()
doesn't have any benefit. In BSD, and Solaris, issetugid() is a kernel syscall
and is safer. In Linux we should use __libc_enable_secure(), which is similar,
but requires packages to be patched. All packages should be searched for the
issetugid() function, and modified to use __libc_enable_secure() instead.