naoki: Get rid of the toolchain.
Big changes in dependency resolution: Naoki does not resolve
dependencies of source packages anymore. It instead goes on
and checks out the available binary packages and tries to
build a one-time customized toolchain.

The toolchain was always a thing that made development slow
and caused lots of issues wo do not have anymore.

This references #53.