Introduce enable/disable option for network zones.
In the past all network zones have been started during the boot processs of the system,
even if they cannot be started because for example the required device is missing.

From now it will be possible to enable or disable network zones for startup. We use the systemd
network.target to set which zones should be started.