Add whitelist of serial USB devices.
If you connect a USB serial adapter to the system, it
will automatically be checked if it is a modem.
We do that by sending AT commands to them and wait for a
response. Some devices that are not a modem may get confused
by that, so we whitelist them here.