Remove support for ISDN
ISDN has never been deployed very much
in the world except Germany and some
other countries. Most of them went away
from ISDN to VoIP for telephony and
DSL or fibre for data connections.

The bandwidth is very small with only
64k per channel and this has only been
implemented for backup purposes if the
primary connection cut off. Nowadays
there is mobile connections with much
more bandwidth available for this.

Nobody is able (due to lack of ISDN)
and willing to support this any more.
Hence the entire support for ISDN
is removed.