Remove pptp
After a rather longer discussion we decided that there are
not many providers any more that use PPTP. The only ones that
we know of are legacy DSL lines in Austria. Most of the newer
DSL providers in Austria use PPPoE now.

Some "privacy" VPN providers allow using PPTP which is just
ridiculous. We do not want to support that.

PPTP is also very rarely used for IPv6 transition, but only
a few providers use that. They always support a different
technology.