Implement identify feature
When executing "network [zone|port|device] XXX identify" the
port, device or all ports of the zone (that support this feature)
will flash for ten seconds.

Ports will also flash when they are attached to or detached from
a zone.

This is supposed to make plugging in a cable into the right
device is easier.

Fixes #10969

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>