header-zone: refactor hook_config_destroy
With the new id function this function gets a zone name and a id,
but hook_config_cmd needs also the name of the hook.

So this function now calls zone_config_get_hook_from_id to get the hook
and calls then hook_config_cmd with the correct argument order.

Signed-off-by: Jonatan Schlag <jonatan.schlag@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>