zone: change edit syntax for config.
The syntax to edit a config is now

network zone upl0 config <id> edit

similar to the syntax if a zone is edited.

The cmd variable is setted to the content of ${1},
because we need the content in this variable if the ${id} is not valid, to print a nice error message.

Signed-off-by: Jonatan Schlag <jonatan.schlag@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>