Add first test for nitsi
This commit introduce NITSI. Nitsi is the "Networking integration test
suite for IPFire". We can test the network code in a virtual environment
on any system.

This test has the only purpose to check if nitsi is working.

For more information about nitsi see the manpages and the git
repository on git.ipfire.org

Signed-off-by: Jonatan Schlag <jonatan.schlag@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>