Add include dir for nitsi test
When we include recipes in our recipe we must be shure in some cases
that the recipe are generated out of a .in file.

All files in the include dir will be generated before every test so we
can be shure that these files are present.

This is useful for recipes like the make-install recipe in this test,
which needs to be generated and will be included in nearly every test.

Signed-off-by: Jonatan Schlag <jonatan.schlag@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>