ip-tunnel: Set TTL to 255 by default
By default, the Linux kernel inherits the TTL of the transported
packet. Usually with BGP, the TTL is deliberately set to 1 or very
low numbers which causes the packet to be dropped after the first
hop.

Since the tunnel should be routed, we set this to a default value
of 255 and ignore the TTL of the encapsulated packet.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>